requirements.txt
allure-pytest==2.8.11
allure-python-commons==2.8.11
Appium-Python-Client==0.50
atomicwrites==1.3.0
attrs==19.3.0
behave==1.2.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
configparser==4.0.2
crayons==0.3.0
et-xmlfile==1.0.1
html-testRunner==1.2.1
idna==2.9
jdcal==1.4.1
Jinja2==2.11.1
MarkupSafe==1.1.1
more-itertools==8.2.0
openpyxl==3.0.3
packaging==20.3
parse==1.15.0
parse-type==0.5.2
path==13.2.0
pluggy==0.13.1
py==1.8.1
py2exe==0.9.2.2
pyparsing==2.4.6
pytest==5.3.5
pytest-html==2.0.1
pytest-metadata==1.8.0
requests==2.23.0
selenium==3.141.0
six==1.14.0
urllib3==1.25.8
wcwidth==0.1.8
webdriver-manager==2.3.0
